+++ spec/installers/heroku_ruby_installer_spec.rb
@@ -1,9 +1,13 @@
 require "spec_helper"
 
 describe LanguagePack::Installers::HerokuRubyInstaller do
-  let(:installer)    { LanguagePack::Installers::HerokuRubyInstaller.new(
-    stack: "cedar-14"
-  ) }
+  let(:installer) {
+    LanguagePack::Installers::HerokuRubyInstaller.new(
+      multi_arch_stacks: [],
+      stack: "cedar-14",
+      arch: nil,
+    )
+  }
   let(:ruby_version) { LanguagePack::RubyVersion.new("ruby-2.3.3") }
 
   describe "#fetch_unpack" do